Retrain model with different environment
The different environment was created to attempt to address an issue related to the image loaded in AWS Lambda. Turned out that the wrong image was being selected (out of the two created), so the new environment was not necessary. The image build command is modified to create a single image file, which works in Lambda.